msm: kgsl: Don't wait for room in context queue...
msm: kgsl: Don't wait for room in context queue when context is invalidated

In a condition where there is no room in the drawcontext queue and the
context is invalidated while sleeping in wait_event_interruptible_timeout,
we are now polling for room after wake up. This would lead to infinite
looping.

Hence, do not wait for room in context queue if context is invalidated.